+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.5] - 2026-01-24
+
+### Fixed
+- Re-sign FerretDB and postgresql-documentdb binaries on macOS after download to fix Gatekeeper code signature issues
+
+### Changed
+- Improved error message when postgresql-documentdb binaries fail to execute due to library loading issues
+
 ## [0.23.4] - 2026-01-23
 
 ### Fixed